Add files via upload
It is a well-known fact that virtually all educational or commercial organizations need to properly record the attendance of their students or employees for effective planning, management, and functioning of the organization. 
In most universities in developing countries, student attendance is usually taken by the old file system approach by calling students' name and using paper sheets, this approach is being used for a long time.
Keeping these issues in mind, we designed and implemented a system to overcome the problems associated with attendance recording.